+++ NewGraph.h
@@ -22,7 +22,7 @@ class Graph
 		Graph();
 		bool validateName(const string);
 		void addPOI(const string);
-		void addRoad(const string, const string, const string, int);
+		void addRoad(const string, const string, const string, double);
 		void getShortestPath(const string);
 	       	void printNodes();
 };
